# docs/swagger/iam.json
{
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Simple IAM verification procedure for OAuth2 authentication cycles. Generates a token for\na valid user with expiration time, and provides functionality for the validation of tokens.",
    "title": "IAM API",
    "version": "APP_VERSION_AUTOGENERATED"
  },
  "host": "localhost:8080",
  "basePath": "/iam/v1",
  "paths": {
    "/health": {
      "get": {
        "produces": [
          "application/json"
        ],
        "summary": "Responds to health checks.",
        "operationId": "health",
        "responses": {
          "200": {
            "description": "health",
            "schema": {
              "$ref": "#/definitions/health"
            }
          },
          "500": {
            "description": ""
          },
          "503": {
            "description": ""
          }
        }
      }
    },
    "/oauth2/identity": {
      "post": {
        "description": "If the token is invalid, an error is returned.",
        "produces": [
          "application/json"
        ],
        "summary": "Responds with the subject embedded in the token claims, if there is one.",
        "operationId": "identity",
        "responses": {
          "200": {
            "description": "tokenIdentity",
            "schema": {
              "$ref": "#/definitions/tokenIdentity"
            }
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          }
        }
      }
    },
    "/oauth2/token": {
      "post": {
        "produces": [
          "application/json"
        ],
        "summary": "Responds with an access token.",
        "operationId": "token",
        "responses": {
          "200": {
            "description": "token",
            "schema": {
              "$ref": "#/definitions/token"
            }
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          }
        }
      }
    },
    "/oauth2/validate": {
      "post": {
        "summary": "Responds with an error if the token is not valid.",
        "operationId": "validate",
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          }
        }
      }
    },
    "/ready": {
      "get": {
        "summary": "Responds with 200 when service is ready to accept requests.",
        "operationId": "ready",
        "responses": {
          "200": {
            "description": ""
          },
          "503": {
            "description": ""
          }
        }
      }
    }
  },
  "definitions": {
    "health": {
      "description": "Health of the service",
      "type": "object",
      "properties": {
        "iam": {
          "type": "string",
          "x-go-name": "IAM"
        },
        "status": {
          "type": "string",
          "enum": [
            "Alive",
            "Degraded",
            "Unavailable"
          ],
          "x-go-name": "Status"
        }
      },
      "x-go-name": "Health",
      "x-go-package": "github.com/ingka-group-digital/iam-proxy/client/health"
    },
    "token": {
      "description": "Token for IAM verification",
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string",
          "x-go-name": "AccessToken"
        },
        "expiresIn": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ExpiresIn"
        },
        "extExpiresIn": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ExtExpiresIn"
        },
        "identityToken": {
          "type": "string",
          "x-go-name": "IdentityToken"
        },
        "tokenType": {
          "type": "string",
          "x-go-name": "TokenType"
        }
      },
      "x-go-name": "Token",
      "x-go-package": "github.com/ingka-group-digital/iam-proxy/client/iam"
    },
    "tokenIdentity": {
      "description": "TokenIdentity for getting the IAM token identity",
      "type": "object",
      "properties": {
        "identity": {
          "type": "string",
          "x-go-name": "Identity"
        }
      },
      "x-go-name": "TokenIdentity",
      "x-go-package": "github.com/ingka-group-digital/iam-proxy/client/iam"
    }
  }
}